refactor : 알림 전역 관리를 위한 SSE 컴포넌트 생성 및 호출 위치 변경
