+++ client/src/App.tsx
@@ -2,6 +2,7 @@ import { RecoilRoot } from 'recoil';
 import { useState } from 'react';
 import { QueryClient, QueryClientProvider } from 'react-query';
 import { Router } from './router/Router';
+import { SSE } from './components/_common/SSE/SSE';
 
 declare module 'react-query/types/react/QueryClientProvider' {
   interface QueryClientProviderProps {
@@ -26,6 +27,7 @@ export const App = () => {
     <div className="flex items-center justify-center w-screen h-screen bg-background-desktop">
       <RecoilRoot>
         <QueryClientProvider client={queryClient}>
+          <SSE />
           <Router />
         </QueryClientProvider>
       </RecoilRoot>
